back_end_zfll/requirements/dev.txt
-r base.txt

# Dev tools
django-debug-toolbar>=4.3,<5.0
django-extensions>=3.2,<4.0
ipython>=8.0,<9.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
faker>=30.0,<31.0

# Code quality
flake8>=7.0,<8.0
black>=24.0,<25.0
isort>=5.13,<6.0
